requirements.txt
geopandas
pandas
numpy
matplotlib
fiona
shapely
pyproj
pyyaml

mapclassify
